[REMOVE] Mocks: json.loads
We are probably going to use the json.loads in a different way. For that reason, we temporarily removed the json.loads mock from the unit tests. We also enhanced the urllib.request.urlopen mock by using logic from the json.loads mock. The samples were adapted to be used by the enhanced mocks.

Related issues: #5