[RESTORE] Demo script
On a previous commit, the demo script src/core/main.py was removed from this repository. However, for usability reasons, it was restored into a new file called download_medical_images.py. To keep it simple, since it is only a demo script, it only accepts a single argument, and it does not download DICOM files on directories with content.

Moreover, while executing the demo script, the following issues were identified and solved:
* Missing urllib.request import at url.py: for some reason, the unit tests were behaving fine while importing urllib only. The url.py now has all the correct imports. Also, the unit tests were adapted to consider this issue.
* Destination path format without the final path delimiter at format.py: this issue was hidden because the path delimiter was passed through the old demo script. In this new demo script, however, the argument is arbitrary, so the path delimiter may not be passed. The format.py now has the path delimiter.

Related issues: #5